requirements.txt
opentelemetry-api
opentelemetry-sdk
otel-extensions>=0.0.4
pytest>=5.0.0